Add files via upload
The gender and age detection ML model uses deep learning algorithms to analyze features such as facial structure, skin color, and other facial attributes to make predictions about a person's gender and age. The model is trained on a large dataset of labeled images that include both males and females from different age groups, and it learns to identify patterns in the data that correlate with gender and age.

The model can be used in a variety of applications, such as security systems that require gender and age identification for access control, marketing research that needs to know the demographics of a particular audience, or entertainment industry that needs to categorize their audience based on age and gender. Additionally, the model can be used in social media and online platforms for demographic analysis and targeted advertising.

Overall, the gender and age detection ML model is a powerful tool for analyzing images and videos, and it has the potential to be a valuable asset for a wide range of applications.